[MSFT] Attempt to fix Windows builds
Change an incorrect usage of "unsigned long" in a definition to match the
"size_t" in the accompanying declaration.

+++ lib/Dialect/MSFT/MSFTOps.cpp
@@ -952,9 +952,7 @@ size_t MSFTModuleExternOp::getNumPorts() {
   return getArgNames().size() + getResultNames().size();
 }
 
-hw::InnerSymAttr MSFTModuleExternOp::getPortSymbolAttr(unsigned long) {
-  return {};
-}
+hw::InnerSymAttr MSFTModuleExternOp::getPortSymbolAttr(size_t) { return {}; }
 
 //===----------------------------------------------------------------------===//
 // OutputOp
